Extract AbstractOrcRecordReader and StreamReader
