+++ presto-orc/src/main/java/com/facebook/presto/orc/reader/StreamReader.java
@@ -13,16 +13,21 @@
  */
 package com.facebook.presto.orc.reader;
 
-import com.facebook.presto.spi.block.Block;
-import com.facebook.presto.spi.type.Type;
+import com.facebook.presto.orc.metadata.ColumnEncoding;
+import com.facebook.presto.orc.stream.InputStreamSources;
 
 import java.io.IOException;
+import java.util.List;
 
-public interface BatchStreamReader
-        extends StreamReader
+public interface StreamReader
 {
-    Block readBlock(Type type)
+    void startStripe(InputStreamSources dictionaryStreamSources, List<ColumnEncoding> encoding)
             throws IOException;
 
-    void prepareNextRead(int batchSize);
+    void startRowGroup(InputStreamSources dataStreamSources)
+            throws IOException;
+
+    void close();
+
+    long getRetainedSizeInBytes();
 }
